Update iter_patches() to return EnrichedFixPatch for better fix placement
